# backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/controller/PartTypeInformationController.java
/*
 * Copyright (c) 2024 Volkswagen AG
 * Copyright (c) 2025 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e.V.
 * (represented by Fraunhofer ISST)
 * Copyright (c) 2024 Contributors to the Eclipse Foundation
 *
 * See the NOTICE file(s) distributed with this work for additional
 * information regarding copyright ownership.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Apache License, Version 2.0 which is available at
 * https://www.apache.org/licenses/LICENSE-2.0.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 */

package org.eclipse.tractusx.puris.backend.masterdata.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import lombok.extern.slf4j.Slf4j;
import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Material;
import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Partner;
import org.eclipse.tractusx.puris.backend.masterdata.logic.adapter.PartTypeInformationSammMapper;
import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.parttypeinformation.PartTypeInformationSAMM;
import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialPartnerRelationService;
import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialService;
import org.eclipse.tractusx.puris.backend.masterdata.logic.service.PartnerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.regex.Pattern;

@RestController
@RequestMapping("parttypeinformation")
@Slf4j
public class PartTypeInformationController {
    static Pattern bpnlPattern = PatternStore.BPNL_PATTERN;
    static Pattern materialNumberPattern = PatternStore.URN_OR_UUID_PATTERN;

    @Autowired
    private PartnerService partnerService;
    @Autowired
    private MaterialService materialService;
    @Autowired
    private MaterialPartnerRelationService mprService;
    @Autowired
    private PartTypeInformationSammMapper sammMapper;

    @RequestMapping(value = "/**")
    @ResponseStatus(HttpStatus.NOT_IMPLEMENTED)
    public ResponseEntity<String> handleNotImplemented() {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

    @Operation(description = "Endpoint that delivers PartTypeInformation of own products to customer partners. " +
        "'globalAssetId' must be set to the global asset id of the party, that receives the request. Please note that the " +
        "SAMMs delivered by this endpoint don't provide partClassification data. " +
        "This endpoint is meant to be accessed by partners via EDC only. ")
    @ApiResponses(value = {
        @ApiResponse(responseCode = "200", description = "Ok"),
        @ApiResponse(responseCode = "400", description = "Invalid request parameters. ", content = @Content),
        @ApiResponse(responseCode = "401", description = "Access forbidden. ", content = @Content),
        @ApiResponse(responseCode = "404", description = "Product not found for given parameters. ", content = @Content),
        @ApiResponse(responseCode = "501", description = "Unsupported representation requested. ", content = @Content)
    })
    @GetMapping("/{globalAssetId}/submodel/{representation}")
    public ResponseEntity<PartTypeInformationSAMM> getMapping(@RequestHeader("edc-bpn") String bpnl,
                                                              @Parameter(description = "The global asset id for the requested part type.")
                                        @PathVariable String globalAssetId,
                                                              @Parameter(description = "Must be set to '$value'") @PathVariable String representation) {
        if (!bpnlPattern.matcher(bpnl).matches() || !materialNumberPattern.matcher(globalAssetId).matches()) {
            return ResponseEntity.badRequest().build();
        }

        if (!"$value".equals(representation)) {
            return ResponseEntity.status(501).build();
        }
        Partner partner = partnerService.findByBpnl(bpnl);
        if (partner == null) {
            return ResponseEntity.status(401).build();
        }
        log.info("{} requests part type information on {}", bpnl, globalAssetId);
        Material material = materialService.findByMaterialNumberCx(globalAssetId);
        if (material == null || !material.isProductFlag()) {
            return ResponseEntity.status(404).build();
        }
        var mpr = mprService.find(material, partner);
        if (mpr == null || !mpr.isPartnerBuysMaterial()) {
            return ResponseEntity.status(404).build();
        }
        var samm = sammMapper.productToSamm(mpr);
        return ResponseEntity.ok(samm);
    }
}