# docs/api/openAPI.yaml
components:
  schemas:
...
    Classification:
      additionalProperties: false
      properties:
        classificationDescription:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationID:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationStandard:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      required:
      - classificationID
      - classificationStandard
      type: object
...
    PartSitesInformationAsPlanned:
      additionalProperties: false
      properties:
        catenaXsiteId:
          maxItems: 50
          pattern: ^BPNS[0-9a-zA-Z]{12}$
          type: string
        function:
          enum:
          - production
          - warehouse
          - spare part warehouse
          maxItems: 50
          type: string
        functionValidFrom:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
        functionValidUntil:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
      required:
      - catenaXsiteId
      - function
      type: object
    PartTypeInformationBody:
      additionalProperties: false
      properties:
        manufacturerPartId:
          maxItems: 50
          type: string
        nameAtManufacturer:
          maxItems: 50
          type: string
        partClassification:
          items:
            $ref: '#/components/schemas/Classification'
          maxItems: 50
          type: array
      required:
      - manufacturerPartId
      - nameAtManufacturer
      type: object
    PartTypeInformationSAMM:
      additionalProperties: false
      properties:
        catenaXId:
          maxItems: 50
          pattern: (^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)|(^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)
          type: string
        partSitesInformationAsPlanned:
          items:
            $ref: '#/components/schemas/PartSitesInformationAsPlanned'
          maxItems: 50
          type: array
          uniqueItems: true
        partTypeInformation:
          $ref: '#/components/schemas/PartTypeInformationBody'
      required:
      - catenaXId
      - partTypeInformation
      type: object
...
  securitySchemes:
    X-API-KEY:
      in: header
      name: X-API-KEY
      type: apiKey
info:
  title: PURIS FOSS Open API
  version: 6.0.0
openapi: 3.1.0
paths:
...
  /parttypeinformation/{globalAssetId}/submodel/{representation}:
    get:
      description: 'Endpoint that delivers PartTypeInformation of own products to
        customer partners. ''globalAssetId'' must be set to the global asset id of
        the party, that receives the request. Please note that the SAMMs delivered
        by this endpoint don''t provide partClassification data. This endpoint is
        meant to be accessed by partners via EDC only. '
      operationId: getMapping
      parameters:
      - in: header
        name: edc-bpn
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: The global asset id for the requested part type.
        in: path
        name: globalAssetId
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: Must be set to '$value'
        in: path
        name: representation
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartTypeInformationSAMM'
                additionalProperties: false
          description: Ok
        '400':
          description: 'Invalid request parameters. '
        '401':
          description: 'Access forbidden. '
        '404':
          description: 'Product not found for given parameters. '
        '501':
          description: 'Unsupported representation requested. '
      tags:
      - part-type-information-controller
...
security:
- X-API-KEY: []
servers:
- description: Generated server url
  url: http://localhost:8081/catena